requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.6.1
cycler==0.10.0
discord.py==1.2.5
idna==2.8
idna-ssl==1.1.0
kiwisolver==1.1.0
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
multidict==4.5.2
nltk==3.4.5
numpy==1.16.2
olefile==0.46
Pillow==5.4.1
psutil==5.6.2
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrez==1.1.0.2
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-aeaweb==0.0.1
six==1.12.0
soupsieve==1.8
textblob==0.15.3
tornado==6.0.2
typing-extensions==3.7.2
urllib3==1.24.1
websockets==6.0
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.3.0
